Changed call to AAH_SERVER based parameterization

+++ script/on_picture_save.sh
@@ -280,13 +280,14 @@ if [ -n "${MQTT_ON}" ] && [ -s "${OUTPUT}" ] && [ -n "${MQTT_HOST}" ]; then
 fi
 
 # force image updates periodically (15 minutes; 1800 seconds)
-TTL=1800
-SECONDS=$(date "+%s")
-DATE=$(/bin/echo "${SECONDS} / ${TTL} * ${TTL}" | bc)
-if [ ! -f "/tmp/images.$DATE.json" ]; then
-  rm -f "/tmp/images".*.json
-  curl "http://www.dcmartin.com/CGI/aah-images.cgi?db=${DEVICE_NAME}" > "/tmp/images.${DATE}.json"
-  jq -c '.' "/tmp/images.${DATE}.json"
+if [ -n "${AAH_SERVER}" ]; then
+  TTL=1800
+  SECONDS=$(date "+%s")
+  DATE=$(/bin/echo "${SECONDS} / ${TTL} * ${TTL}" | bc)
+  if [ ! -f "/tmp/images.$DATE.json" ]; then
+    rm -f "/tmp/images".*.json
+    curl "http://${AAH_SERVER}/CGI/aah-images.cgi?db=${DEVICE_NAME}" > "/tmp/images.${DATE}.json"
+    jq -c '.' "/tmp/images.${DATE}.json"
+  fi
 fi
-
 echo "+++ END: $0: $*" $(date) >&2
